+++ crates/xilem_svg/web_examples/svgtoy/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "svgtoy"
+version = "0.1.0"
+publish = false
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
+console_error_panic_hook = "0.1"
+wasm-bindgen = "0.2.87"
+web-sys = "0.3.64"
+xilem_svg = { path = "../.." }
