+++ crates/xilem_svg/web_examples/svgtoy/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<head>
+<style>
+    .gray {
+        background-color: lightgrey;
+    }
+    .green {
+        background-color: lightgreen;
+    }
+    rect.red {
+        fill: #e00;
+    }
+    line {
+        stroke: #444;
+    }
+</style>
+</head>
+<body>
+</body>
